Preliminary collision detection

+++ TronLightCycles/Particle.pde
@@ -1,7 +1,7 @@
 class Particle{
-  int px,py;
-  int pr, pg, pb;
- Particle(int x, int y, int r, int g, int b){
+  float px,py;
+  float pr, pg, pb;
+ Particle(float x, float y, float r, int g, int b){
   px = x;
   py = y;  
   pr = r;
